First attempt for prototyping table of quotes
fixed imports

wip: fixing ank imports in documentation

documentation ank fixes

Figured out how to add quotes readme separate from index and added some available examples

wip: metaextension refactor part 1

Attempted file restructure to reflect the hierarchy of the PSI system in Kotlin as well as a major refactor to bring all meta extensions to the root of the quote system so they can be made as a quasiquote table in autogeneration. Some further refactor to replace any existing scopes as well

left out one import

smh left out another import

moved typereference to elements

Code cleanup

some code cleanup and more attempts with ank for dokka with no avail

rebase cleanup

removed Ank for now for Paco to look further into, but added the examples for meta extension declarations

code cleanup

import fix

Totally redid metaexpressions and how we do documentation for the elmenets

Update compiler-plugin/src/main/kotlin/arrow/meta/quotes/MetaExtensions.kt

Added some clarity for test plugins
