+++ compiler-plugin/src/main/kotlin/arrow/meta/phases/analysis/ElementScope.kt
@@ -1,29 +1,36 @@
 package arrow.meta.phases.analysis
 
-import arrow.meta.quotes.AnnotatedExpression
-import arrow.meta.quotes.BlockExpression
-import arrow.meta.quotes.CatchClause
-import arrow.meta.quotes.ClassDeclaration
-import arrow.meta.quotes.FinallySection
-import arrow.meta.quotes.ForExpression
-import arrow.meta.quotes.IfExpression
-import arrow.meta.quotes.ImportDirective
-import arrow.meta.quotes.IsExpression
-import arrow.meta.quotes.ModifierList
-import arrow.meta.quotes.NamedFunction
-import arrow.meta.quotes.ObjectDeclaration
-import arrow.meta.quotes.ParameterList
-import arrow.meta.quotes.Parameter
-import arrow.meta.quotes.ReturnExpression
 import arrow.meta.quotes.Scope
-import arrow.meta.quotes.ThrowExpression
-import arrow.meta.quotes.TryExpression
-import arrow.meta.quotes.TypeReference
-import arrow.meta.quotes.ValueArgument
-import arrow.meta.quotes.WhenCondition
-import arrow.meta.quotes.WhenEntry
-import arrow.meta.quotes.WhenExpression
-import arrow.meta.quotes.WhileExpression
+import arrow.meta.quotes.classorobject.ClassDeclaration
+import arrow.meta.quotes.classorobject.ObjectDeclaration
+import arrow.meta.quotes.declaration.DestructuringDeclaration
+import arrow.meta.quotes.element.CatchClause
+import arrow.meta.quotes.element.FinallySection
+import arrow.meta.quotes.element.ImportDirective
+import arrow.meta.quotes.element.ParameterList
+import arrow.meta.quotes.element.ValueArgument
+import arrow.meta.quotes.element.WhenCondition
+import arrow.meta.quotes.element.WhenEntry
+import arrow.meta.quotes.expression.AnnotatedExpression
+import arrow.meta.quotes.expression.BinaryExpression
+import arrow.meta.quotes.expression.BlockExpression
+import arrow.meta.quotes.expression.IfExpression
+import arrow.meta.quotes.expression.IsExpression
+import arrow.meta.quotes.expression.LambdaExpression
+import arrow.meta.quotes.expression.ThrowExpression
+import arrow.meta.quotes.expression.TryExpression
+import arrow.meta.quotes.expression.WhenExpression
+import arrow.meta.quotes.expression.expressionwithlabel.PropertyAccessor
+import arrow.meta.quotes.expression.expressionwithlabel.ReturnExpression
+import arrow.meta.quotes.expression.loopexpression.ForExpression
+import arrow.meta.quotes.expression.loopexpression.WhileExpression
+import arrow.meta.quotes.modifierlist.ModifierList
+import arrow.meta.quotes.modifierlist.TypeReference
+import arrow.meta.quotes.nameddeclaration.notstubbed.FunctionLiteral
+import arrow.meta.quotes.nameddeclaration.stub.Parameter
+import arrow.meta.quotes.nameddeclaration.stub.typeparameterlistowner.NamedFunction
+import arrow.meta.quotes.nameddeclaration.stub.typeparameterlistowner.Property
+import arrow.meta.quotes.nameddeclaration.stub.typeparameterlistowner.TypeAlias
 import org.jetbrains.kotlin.com.intellij.openapi.project.Project
 import org.jetbrains.kotlin.com.intellij.psi.PsiComment
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
@@ -36,7 +43,6 @@ import org.jetbrains.kotlin.psi.KtCallableReferenceExpression
 import org.jetbrains.kotlin.psi.KtClassBody
 import org.jetbrains.kotlin.psi.KtConstructorDelegationCall
 import org.jetbrains.kotlin.psi.KtDeclaration
-import org.jetbrains.kotlin.psi.KtDestructuringDeclaration
 import org.jetbrains.kotlin.psi.KtDotQualifiedExpression
 import org.jetbrains.kotlin.psi.KtEnumEntry
 import org.jetbrains.kotlin.psi.KtExpression
@@ -49,7 +55,6 @@ import org.jetbrains.kotlin.psi.KtLiteralStringTemplateEntry
 import org.jetbrains.kotlin.psi.KtPackageDirective
 import org.jetbrains.kotlin.psi.KtPrimaryConstructor
 import org.jetbrains.kotlin.psi.KtProperty
-import org.jetbrains.kotlin.psi.KtPropertyAccessor
 import org.jetbrains.kotlin.psi.KtPropertyDelegate
 import org.jetbrains.kotlin.psi.KtSecondaryConstructor
 import org.jetbrains.kotlin.psi.KtSimpleNameExpression
@@ -59,7 +64,6 @@ import org.jetbrains.kotlin.psi.KtStringTemplateExpression
 import org.jetbrains.kotlin.psi.KtSuperTypeCallEntry
 import org.jetbrains.kotlin.psi.KtSuperTypeEntry
 import org.jetbrains.kotlin.psi.KtThisExpression
-import org.jetbrains.kotlin.psi.KtTypeAlias
 import org.jetbrains.kotlin.psi.KtTypeArgumentList
 import org.jetbrains.kotlin.psi.KtTypeCodeFragment
 import org.jetbrains.kotlin.psi.KtTypeElement
@@ -106,13 +110,13 @@ interface ElementScope {
     name: String,
     typeParameters: List<String>,
     typeElement: KtTypeElement
-  ): Scope<KtTypeAlias>
+  ): TypeAlias
   
   fun typeAlias(
     name: String,
     typeParameters: List<String>,
     body: String
-  ): Scope<KtTypeAlias>
+  ): TypeAlias
   
   val star: PsiElement
   
@@ -165,15 +169,15 @@ interface ElementScope {
     isVar: Boolean
   ): Scope<KtProperty>
   
-  val String.property: Scope<KtProperty>
+  val String.property: Property
   
-  fun propertyGetter(expression: KtExpression): Scope<KtPropertyAccessor>
+  fun propertyGetter(expression: KtExpression): PropertyAccessor
   
-  fun propertySetter(expression: KtExpression): Scope<KtPropertyAccessor>
+  fun propertySetter(expression: KtExpression): PropertyAccessor
   
   fun propertyDelegate(expression: KtExpression): Scope<KtPropertyDelegate>
   
-  val String.destructuringDeclaration: Scope<KtDestructuringDeclaration>
+  val String.destructuringDeclaration: DestructuringDeclaration
   
   val String.destructuringParameter: Parameter
   
@@ -190,6 +194,8 @@ interface ElementScope {
   val String.identifier: PsiElement
   
   val String.function: NamedFunction
+
+  val String.binaryExpression: BinaryExpression
   
   val String.callableReferenceExpression: Scope<KtCallableReferenceExpression>
   
@@ -312,6 +318,12 @@ interface ElementScope {
   val String.`return`: ReturnExpression
 
   val String.annotatedExpression: AnnotatedExpression
+
+  val String.lambdaExpression: LambdaExpression
+
+  val String.importDirective: ImportDirective
+
+  val String.functionLiteral: FunctionLiteral
 
   /**
    * Creates an expression that has reference to its context
