+++ compiler-plugin/src/main/kotlin/arrow/meta/plugins/helloWorld/DummyPlugin.kt
@@ -4,7 +4,7 @@ import arrow.meta.Meta
 import arrow.meta.Plugin
 import arrow.meta.invoke
 import arrow.meta.quotes.Transform
-import arrow.meta.quotes.nameddeclaration.typeparameterlistowner.namedFunction
+import arrow.meta.quotes.namedFunction
 
 /**
  * The following example shows a Hello World Compiler Plugin.
