+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/classorobject/ObjectDeclaration.kt
@@ -1,23 +1,7 @@
 package arrow.meta.quotes.classorobject
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
-import arrow.meta.quotes.Transform
 import arrow.meta.quotes.Scope
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.psi.KtObjectDeclaration
-
-/**
- * A template destructuring [Scope] for a [KtObjectDeclaration]
- *
- * @param match designed to to feed in any kind of [KtObjectDeclaration] predicate returning a [Boolean]
- * @param map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.objectDeclaration(
-  match: KtObjectDeclaration.() -> Boolean,
-  map: ObjectDeclaration.(KtObjectDeclaration) -> Transform<KtObjectDeclaration>
-): ExtensionPhase =
-  quote(match, map) { ObjectDeclaration(it) }
 
 /**
  * A template destructuring [Scope] for a [KtObjectDeclaration]
