+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/declaration/DestructuringDeclaration.kt
@@ -1,20 +1,17 @@
 package arrow.meta.quotes.declaration
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.psi.KtDestructuringDeclaration
 import org.jetbrains.kotlin.psi.KtDestructuringDeclarationEntry
 import org.jetbrains.kotlin.psi.KtExpression
 
 /**
- * A [KtDestructuringDeclaration] [Quote] with a custom template destructuring [DestructuringDeclaration]. See below:
+ * <code>```""" $`(vars)` = $initializer """.destructuringDeclaration```</code>
  *
- * ```kotlin:ank:silent
+ * A template destructuring [Scope] for a [KtDestructuringDeclaration]. See below:
+ *
+ *  ```kotlin:ank:silent
  * import arrow.meta.Meta
  * import arrow.meta.Plugin
  * import arrow.meta.invoke
@@ -33,19 +30,7 @@ import org.jetbrains.kotlin.psi.KtExpression
  *     }
  *    )
  *   }
- *```
- *
- * @param match designed to to feed in any kind of [KtDestructuringDeclaration] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.destructuringDeclaration(
-  match: KtDestructuringDeclaration.() -> Boolean,
-  map: DestructuringDeclaration.(KtDestructuringDeclaration) -> Transform<KtDestructuringDeclaration>
-): ExtensionPhase =
-  quote(match, map) { DestructuringDeclaration(it) }
-
-/**
- * A template destructuring [Scope] for a [KtDestructuringDeclaration]
+ * ```
  */
 class DestructuringDeclaration(
   override val value: KtDestructuringDeclaration?,
