+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/element/CatchClause.kt
@@ -1,18 +1,15 @@
 package arrow.meta.quotes.element
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.nameddeclaration.ParameterScope
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
+import arrow.meta.quotes.nameddeclaration.stub.Parameter
 import org.jetbrains.kotlin.psi.KtCatchClause
 import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.kotlin.psi.KtParameterList
 
 /**
- * A [KtCatchClause] [Quote] with a custom template destructuring [CatchClause]. See below:
+ * <code>```""" catch $`(parameter)` $`{ body }` """.catchClause```</code>
+ *
+ * A template destructuring [Scope] for a [KtCatchClause]. See below:
  *
  * ```kotlin:ank:silent
  * import arrow.meta.Meta
@@ -34,18 +31,6 @@ import org.jetbrains.kotlin.psi.KtParameterList
  *    )
  *   }
  *```
- *
- * @param match designed to to feed in any kind of [KtCatchClause] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.catchClause(
-  match: KtCatchClause.() -> Boolean,
-  map: CatchClause.(KtCatchClause) -> Transform<KtCatchClause>
-): ExtensionPhase =
-  quote(match, map) { CatchClause(it) }
-
-/**
- * A template destructuring [Scope] for a [KtCatchClause]
  */
 class CatchClause(
   override val value: KtCatchClause?,
