+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/element/FinallySection.kt
@@ -1,16 +1,13 @@
 package arrow.meta.quotes.element
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.psi.KtBlockExpression
 import org.jetbrains.kotlin.psi.KtFinallySection
 
 /**
- * A [KtFinallySection] [Quote] with a custom template destructuring [FinallySection]. See below:
+ * <code>""" $finally """.finally</code>
+ *
+ * A template destructuring [Scope] for a [KtFinallySection]. See below:
  *
  * ```kotlin:ank:silent
  * import arrow.meta.Meta
@@ -32,18 +29,6 @@ import org.jetbrains.kotlin.psi.KtFinallySection
  *    )
  *   }
  *```
- *
- * @param match designed to to feed in any kind of [KtFinallySection] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.finallySection(
-  match: KtFinallySection.() -> Boolean,
-  map: FinallySection.(KtFinallySection) -> Transform<KtFinallySection>
-): ExtensionPhase =
-  quote(match, map) { FinallySection(it) }
-
-/**
- * A template destructuring [Scope] for a [KtFinallySection]
  */
 class FinallySection(
   override val value: KtFinallySection?,
