+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/element/ValueArgument.kt
@@ -1,0 +1,15 @@
+package arrow.meta.quotes.element
+
+import arrow.meta.quotes.Scope
+import org.jetbrains.kotlin.psi.KtExpression
+import org.jetbrains.kotlin.psi.KtValueArgument
+import org.jetbrains.kotlin.psi.KtValueArgumentName
+
+/**
+ * A template destructuring [Scope] for a [ValueArgument]
+ */
+class ValueArgument(
+  override val value: KtValueArgument,
+  val argumentExpression: Scope<KtExpression>? = Scope(value.getArgumentExpression()),
+  val argumentName: Scope<KtValueArgumentName>? = Scope(value.getArgumentName()) // TODO KtValueArgumentName scope and quote template
+): Scope<KtValueArgument>(value)
