+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/element/WhenEntry.kt
@@ -1,18 +1,15 @@
 package arrow.meta.quotes.element
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.kotlin.psi.KtWhenCondition
 import org.jetbrains.kotlin.psi.KtWhenEntry
 
 /**
- * A [KtWhenEntry] [Quote] with a custom template destructuring [WhenEntry]. See below:
+ * <code>""" $conditions -> $expression """.whenEntry</code>
+ *
+ * A template destructuring [Scope] for a [KtWhenEntry]. See below:
  *
  * ```kotlin:ank:silent
  * import arrow.meta.Meta
@@ -34,18 +31,6 @@ import org.jetbrains.kotlin.psi.KtWhenEntry
  *    )
  *   }
  * ```
- *
- * @param match designed to to feed in any kind of [KtWhenEntry] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.whenEntry(
-  match: KtWhenEntry.() -> Boolean,
-  map: WhenEntry.(KtWhenEntry) -> Transform<KtWhenEntry>
-): ExtensionPhase =
-  quote(match, map) { WhenEntry(it) }
-
-/**
- * A template destructuring [Scope] for a [KtWhenEntry]
  */
 class WhenEntry(
   override val value: KtWhenEntry?,
