+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/expression/BlockExpression.kt
@@ -1,17 +1,14 @@
 package arrow.meta.quotes.expression
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.psi.KtBlockExpression
 import org.jetbrains.kotlin.psi.KtExpression
 
 /**
- * A [KtBlockExpression] [Quote] with a custom template destructuring [BlockExpression]. See below:
+ * <code>""" $`@annotations` $expression """.annotatedExpression</code>
+ *
+ * A template destructuring [Scope] for a [KtBlockExpression]. See below:
  *
  * ```kotlin:ank:silent
  * import arrow.meta.Meta
@@ -32,19 +29,6 @@ import org.jetbrains.kotlin.psi.KtExpression
  *     }
  *    )
  *   }
- *```
- *
- * @param match designed to to feed in any kind of [KtBlockExpression] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.blockExpression(
-  match: KtBlockExpression.() -> Boolean,
-  map: BlockExpression.(KtBlockExpression) -> Transform<KtBlockExpression>
-): ExtensionPhase =
-  quote(match, map) { BlockExpression(it) }
-
-/**
- * A template destructuring [Scope] for a [KtBlockExpression]
  */
 class BlockExpression(
   override val value: KtBlockExpression?,
