+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/expression/TryExpression.kt
@@ -1,18 +1,16 @@
 package arrow.meta.quotes.expression
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.psi.KtBlockExpression
 import org.jetbrains.kotlin.psi.KtCatchClause
 import org.jetbrains.kotlin.psi.KtFinallySection
 import org.jetbrains.kotlin.psi.KtTryExpression
 
 /**
- * A [KtTryExpression] [Quote] with a custom template destructuring [TryExpression]. See below:
+ * <code>""" try $tryBlock $catch $finally """.tryExpression</code>
+ *
+ * A template destructuring [Scope] for a [KtTryExpression]. See below:
  *
  * ```kotlin:ank:silent
  * import arrow.meta.Meta
@@ -30,22 +28,10 @@ import org.jetbrains.kotlin.psi.KtTryExpression
  *       replacing = e,
  *       newDeclaration = """ try $tryBlock $catch $finally """.`try`
  *      )
-*      }
-*     )
-*    }
+ *      }
+ *     )
+ *    }
  * ```
- *
- * @param match designed to to feed in any kind of [KtTryExpression] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.tryExpression(
-  match: KtTryExpression.() -> Boolean,
-  map: TryExpression.(KtTryExpression) -> Transform<KtTryExpression>
-): ExtensionPhase =
-  quote(match, map) { TryExpression(it) }
-
-/**
- * A template destructuring [Scope] for a [KtTryExpression]
  */
 class TryExpression(
   override val value: KtTryExpression?,
