+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/expression/WhenExpression.kt
@@ -1,19 +1,16 @@
 package arrow.meta.quotes.expression
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.nameddeclaration.typeparameterlistowner.PropertyScope
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
+import arrow.meta.quotes.nameddeclaration.stub.typeparameterlistowner.Property
 import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.kotlin.psi.KtWhenEntry
 import org.jetbrains.kotlin.psi.KtWhenExpression
 
 /**
- * A [KtWhenExpression] [Quote] with a custom template destructuring [WhenExpression].  See below:
+ * <code>""" $condition """.whenCondition</code>
+ *
+ * A template destructuring [Scope] for a [KtWhenExpression]. See below:
  *
  * ```kotlin:ank:silent
  * import arrow.meta.Meta
@@ -37,18 +34,6 @@ import org.jetbrains.kotlin.psi.KtWhenExpression
  *        )})
  *     }
  * ```
- *
- * @param match designed to to feed in any kind of [KtWhenExpression] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.whenExpression(
-  match: KtWhenExpression.() -> Boolean,
-  map: WhenExpression.(KtWhenExpression) -> Transform<KtWhenExpression>
-): ExtensionPhase =
-  quote(match, map) { WhenExpression(it) }
-
-/**
- * A template destructuring [Scope] for a [KtWhenExpression]
  */
 class WhenExpression(
   override val value: KtWhenExpression?,
