+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/expression/loopexpression/LoopExpression.kt
@@ -1,4 +1,4 @@
-package arrow.meta.quotes.expression
+package arrow.meta.quotes.expression.loopexpression
 
 import arrow.meta.quotes.Scope
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
@@ -8,7 +8,7 @@ import org.jetbrains.kotlin.psi.KtLoopExpression
 /**
  * A base template destructuring [Scope] for all [KtLoopExpression] AST elements
  */
-open class LoopExpressionScope<out T : KtLoopExpression>(
+open class LoopExpression<out T : KtLoopExpression>(
   override val value: T,
   val leftParenthesis: PsiElement? = value.leftParenthesis,
   val rightParenthesis: PsiElement? = value.rightParenthesis,
