+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/modifierlist/ModifierList.kt
@@ -1,21 +1,18 @@
 package arrow.meta.quotes.modifierlist
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.psi.KtAnnotationEntry
 import org.jetbrains.kotlin.psi.KtModifierList
 import org.jetbrains.kotlin.psi.psiUtil.visibilityModifier
 
 /**
- * A [KtModifierList] [Quote] with a custom template destructuring [ModifierList].  See below:
+ * <code>""" $`@annotations` $modifier value """.`modifierList`</code>
  *
- *```kotlin:ank:silent
+ * A template destructuring [Scope] for a [KtModifierList]. See below:
+ *
+ * ```kotlin:ank:silent
  * import arrow.meta.Meta
  * import arrow.meta.Plugin
  * import arrow.meta.invoke
@@ -35,18 +32,6 @@ import org.jetbrains.kotlin.psi.psiUtil.visibilityModifier
  *   )
  *  }
  *```
- *
- * @param match designed to to feed in any kind of [KtModifierList] predicate returning a [Boolean]
- * @param map map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.modifierList(
-  match: KtModifierList.() -> Boolean,
-  map: ModifierList.(KtModifierList) -> Transform<KtModifierList>
-): ExtensionPhase =
-  quote(match, map) { ModifierList(it) }
-
-/**
- * A template destructuring [Scope] for a [KtModifierList]
  */
 class ModifierList(
   override val value: KtModifierList?,
