+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/nameddeclaration/stub/Parameter.kt
@@ -1,28 +1,15 @@
-package arrow.meta.quotes.nameddeclaration
+package arrow.meta.quotes.nameddeclaration.stub
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.psi.*
+import org.jetbrains.kotlin.psi.KtAnnotationEntry
+import org.jetbrains.kotlin.psi.KtExpression
+import org.jetbrains.kotlin.psi.KtParameter
+import org.jetbrains.kotlin.psi.KtTypeParameter
+import org.jetbrains.kotlin.psi.KtTypeReference
 import org.jetbrains.kotlin.psi.psiUtil.modalityModifierType
 import org.jetbrains.kotlin.psi.psiUtil.visibilityModifierType
-
-/**
- * A [KtParameter] [Quote] with a custom template destructuring [Parameter]
- *
- * @param match designed to to feed in any kind of [KtParameter] predicate returning a [Boolean]
- * @param map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.parameter(
-  match: KtParameter.() -> Boolean,
-  map: Parameter.(KtParameter) -> Transform<KtParameter>
-) : ExtensionPhase =
-  quote(match, map) { Parameter(it) }
 
 /**
  * A template destructuring [Scope] for a [KtParameter]
