+++ compiler-plugin/src/main/kotlin/arrow/meta/quotes/nameddeclaration/stub/typeparameterlistowner/NamedFunction.kt
@@ -1,14 +1,9 @@
-package arrow.meta.quotes.nameddeclaration.typeparameterlistowner
+package arrow.meta.quotes.nameddeclaration.stub.typeparameterlistowner
 
-import arrow.meta.Meta
-import arrow.meta.phases.ExtensionPhase
 import arrow.meta.phases.analysis.body
 import arrow.meta.phases.analysis.bodySourceAsExpression
 import arrow.meta.quotes.Scope
 import arrow.meta.quotes.ScopedList
-import arrow.meta.quotes.Transform
-import arrow.meta.quotes.Quote
-import arrow.meta.quotes.quote
 import org.jetbrains.kotlin.name.Name
 import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.kotlin.psi.KtNamedFunction
@@ -17,18 +12,6 @@ import org.jetbrains.kotlin.psi.KtTypeParameter
 import org.jetbrains.kotlin.psi.KtTypeReference
 import org.jetbrains.kotlin.psi.psiUtil.modalityModifierType
 import org.jetbrains.kotlin.psi.psiUtil.visibilityModifierType
-
-/**
- * A [KtNamedFunction] [Quote] with a custom template destructuring [NamedFunction]
- *
- * @param match designed to to feed in any kind of [KtNamedFunction] predicate returning a [Boolean]
- * @param map a function that maps over the resulting action from matching on the transformation at the PSI level.
- */
-fun Meta.namedFunction(
-  match: KtNamedFunction.() -> Boolean,
-  map: NamedFunction.(KtNamedFunction) -> Transform<KtNamedFunction>
-): ExtensionPhase =
-  quote(match, map) { NamedFunction(it) }
 
 /**
  * A template destructuring [Scope] for a [KtNamedFunction]
