+++ docs/_data/sidebar.yml
@@ -21,22 +21,22 @@ options:
         url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/
 
       - title: Quotes
-        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/quotes.html
+        url: /quotes
 
-      - title: classOrObject
-        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/class-or-object.html
+      - title: class
+        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/classorobject/class.html
 
-      - title: func
-        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/func.html
+      - title: namedFunction
+        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/nameddeclaration/typeparameterlistowner/named-function.html
 
       - title: property
-        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/property.html
+        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/nameddeclaration/typeparameterlistowner/property.html
 
       - title: propertyAccessor
-        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/property-accessor.html
+        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/expression/expressionwithlabelproperty-accessor.html
 
       - title: typeAlias
-        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/type-alias.html
+        url: /docs/apidocs/compiler-plugin/arrow.meta.quotes/nameddeclaration/typearameterlistowner/type-alias.html
 
   - title: Compiler API
