+++ gradle.properties
@@ -9,7 +9,7 @@ KOTLIN_IDEA_VERSION=1.3.50-release-IJ2019.2-1
 JVM_TARGET_VERSION=1.8
 KOTLIN_TEST_VERSION=3.3.1
 ASSERTJ_VERSION=3.13.2
-DOKKA_VERSION=0.9.18
+DOKKA_VERSION=0.10.0
 CLASS_GRAPH_VERSION=4.8.47
 JUNIT_VINTAGE_VERSION=5.4.0
 KOTLIN_COMPILE_TESTING=1.2.51
